ISSUE #1168 - Renamed function in backedn, moved getModelsData to modelsettings
